+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Add a top-level `lastBy()` function that converts an `Iterable` to a `Map` by
   grouping its elements using a function, keeping the last element for each
-  computed key.
+  computed key. Also available as an extension method on `Iterable`.
 
 ## 1.16.0